Wait for every thread before exiting
Even though exit() should be thread safe, we still get some errors
like 'heap use after free' or 'double free', so we do it the clean
way. We call thread-specific atexit handlers which are going to
release the lock if any, set a global thread error and call
ft_thread_exit (which is a wrapper for pthread_exit).

So now we wait for every worker thread to see that it has to exit
in the cleanup atexit handler. It takes a bit of time but it's not
a problem. Every thread finishes what it is doing before exiting.

We could use signals if we wanted to hurry the process, but it is
not mandatory and looks like it could complicate everything for
something not really useful or required.

TODO: Maybe find a way to not print the exit message right away.
Like storing it into a global and using it on the last exit call.