Get destination with getaddrinfo
We use getaddrinfo instead of gethostbyname to handle IPv6 addresses
too. Right now the get_destinfo function is not called, although it has
been tested, because for tests purposes the scans need to be launched on
hosts that dont exist. Of course if the destination is invalid in normal
circumstances we do not scan it.

This is just a way of testing a lot addresses easily. When networks will
be handled this wont be necessary anymore.