Update TODO once again with our latest findings
This is pretty bad. But at least now I know what the problem is. It
would seem that pcap_breakloop just does not work across threads as the
documentation actually states. This means we have no way of stopping
threads that do not receive answers to their requests.

The ft_nmap structure could be drastically changed but this would
require a lot of work. It would also mean changing the way we handle
scans, tasks and jobs since jobs will have to be done sequentially
because only one instance of pcap_open_live will be running at any time.
This means only one filter at a time and one interface. So there would
be cases where it would not be possible to run scans from the next job.
Like if we have a job on localhost fallowed by one on goole.com. The
first one would require listening on the loopback interface and the
second one on the default interface. Bad, but not really bad.

Also the way single-threaded scans work will have to be reimagined
because obviously both sending and reading would have to be performed
in the same thread.

That's a lot, me who thought I was almost done with ft_nmap x)