Remove next_job call from start_workers
Now this call is only done inside the worker loop. This is better as we
no longer need to use a 'do while' loop and make the code more modular.

Now everything job related happens in a worker. This accelerates the
thread creation loop that is no longer interrupted by mutex locks.

The only drawback is that we can create more threads than we need but
this is not a problem as thread creation is really fast and useless
workers will just kill themselves as soon as they are born.