Read on "any" interface
Also support multiple link layer headers. Use pcap_datalink for this.
It's not in the subject but it's the right way to do this. I'm not going
to hard-code a header type when it can potentially change and the man
explicitly states that it should be checked for. Anyway this is what
allows to read from the Any interface since it's not guaranteed to use
any one link layer header type. So this could be a bonus.