Design the retarded waiting system (TM)
With sig_atomic_t end and pending_tasks variables and the two priority
mutexes the worker threads should wait on a queue at the low level lock
while the first woker thread arrived there should infinitely loop on
these conditions (therefore waiting for a task to be pushed to the
worker queue). During the loop the lock is not even reached if there is
not task because of the pending_tasks variable. This means there cannot
be thread starvation in this function by the waiting thread since the
"write" mutex is not locked so the push function can instantly add new
tasks without having to wait for the high level lock. That's pretty
good, even for the retarded waiting system.