Check if incoming packet ports match a probe
This is because the incoming packets are going to be checked with port
ranges for an easier and shorter pcap filter program to compile and
apply. This means that some packets could pass this filter if both the
source port and the destination are in the port ranges while not being
valid together.