Delete done boolean from probe
Replace it with retry count down. The reply value is initialized at
MAX_REPLY + 1, which means the probes are sent at least MAX_REPLY times
before timeout is set. Although probes can be sent only MAX_REPLY - 1
times if the alarm_handler call happens before a probe is sent but after
it is initialized (so between NEW_HOST and PROBE tasks). This is
unlikely and not a big problem anyway (timeout is set to 4 and even 3 is
large enough).