Make it possible to use wifi interfaces
Now if the default ethernet interface is not available ft_nmap will
switch to the next valid interface. This is done by checking interface
for the IFF_LOWER_UP and IFF_DORMANT linux flags. This only keeps valid
interfaces. So it can switch to wifi if the interface is available and
has either IPV4 or IPV6 as requested by the user.