fix: udp port parsing (reset parse_comma_list)
This is needed because parse_comma_list is a container function with a
pointer to current string stored inside. Because we use a static buffer
it still has the same address when we pass a new list to it so it's not
reset. It has to be done by passing NULL to it before parsing a new
port list.