test: add failing case
I mean it failed once there (with the hosts file). This could be because
of a problem of reply update from the preceding scan but I have never
seen the problem again.

It showed almost all ports as filtered on the framasoft scan, when they
are usually all closed. Some ports were closed as expected. But others
were not showed as open when the presumably should have been and some
where unfiltered (with ACK scan), when there is 0 unfiltered in the
original scan (except for the open ports).