feat: use probe_all task instead of probe
This sends all the probes at the same time in a loop. This is for test
purposes right now. Not sure this is going to make it on the master
branch. It needs to be further tested. Also it needs to work in
monothreaded mode too (which it does not right now).