fix: move breakloop call back to alarm_handler
The pcap_breakloop function was called in the reply task. This was not
working because in multithread runs the reply task is done in threads.
The pcap_breakloop function only works in the thread of the pcap loop.

So the alarm_handler is the only way to stop the listen task.