fix: use task_match exec_time in get_task
The entire list is compared to the same timestamp, which is the output
of gettimeofday, so the current unix timestamp. It means that the tasks
will be executed in the order of the task list and that gettimeofday is
not called for each task. That makes one less TODO to worry about :)