test: non-blocking listener with connect/recv
This looks like it might works. The listener as is gets all the incoming
traffic from a given host. Even if we cannot set a specific port for
listening this is not a big deal because as is we can listen to multiple
ports at the same time, which is pretty good.

Right now it only works with TCP packets. If it can work with UDP and
ICMP, this might be the solution to get a non-blocking listening
function and have real sub-second timeouts in monothread runs.