test: fixed ipv6 icmp
The filter was working just fine (it's a pretty simple filter). It was
the ICMP option which was badly set. I basically did something that
works in JS not in C -_-

I mean the fallback syntax is cool and all, but that's not how it works.
Honestly, now that I think about it, it probably does not work in JS
either. This is a fallback syntax that only exists in my mind.