feat: replace pcap library by custom bpf filters
This makes polling possible instead of using pcap_dispatch and blocking
the main thread. Since pcap_dispatch and filters are useless now, the
library has been removed from the projet. The pcap filter could have
stayed for multithreaded mode, but it's simpler to just use bpf and poll
everywhere. It seems to be working at least with SYN scan and localhost.

But there big failures when scanning remote hosts because of
unpredictable packet size. The pcap implementation was giving the entire
packet to the pcap_handler, now it's capped at RAW_DATA_MAXSIZE which is
problem when parsing the packet. The input buffer should either be
extended or packet parsing should work with truncated packets.