fix: make sure no unfiltered packet passes through
A filter that just drop every packet is set just after the socket is
opened. Then we loop on recv with MSG_DONTWAIT set on until we get an
error (which we assume to be EAGAIN or EWOULDBLOCK). This empties the
random packets that arrived before the zero packet filter was set. Then
everythin is blocked until we set a real filter. Clean, easy, cool ;)