fix: fix UDP scan infinite loop the right way
Now every UDP scan is done one after the other like any other scan. This
is the right way because now the probe task function simply loops with a
modulo operation on the tries value. This means that there never is
multiple pending probe tasks for the same scan_job, only one. This also
means that the UDP ports are not flood with probes. So thats a nice
bonus. Anyway, it fits better and it fixes the problem. Cool :)