fix: start by flushing tasks in NEW_HOST task
This drastically reduces the remaining tasks problem. It sill happens as
I expected but it's pretty rare. I have to loop a lot before seeing it
again. Although I did not try with other hosts, so maybe localhost is a
special case here (probably actually). But even then, it does not matter
the problem still has to be fixed, rare or not.