fix: big memory leak from set_host_job_data
The port_jobs array was reset to 0 each time the function was called.
Which means the allocated probe array pointer on each job was erased
without its content being freed. Then it was allocated again, which was
totally useless.

Now it would not have shown up in valgrind because of how my libft
handles allocated memory, but it was still pretty bad. Glad it's fixed.