fix: wait for running tasks in new_host function
A counter of runnning tasks has been set up so that ft_nmap can wait for
them to end before resetting the host_job. This should fix all the
problems related to this datarace between the new_host function and the
probe/reply tasks running in threads.

Now there is loop in the new_host function that waits for running_tasks
to be equal to one so that only the NEW_HOST task is currently running.
And even if we are really really unlucky (and I'm not even sure this is
possible) and the flushing of the task list and the wait loop somehow
'misses' a task (if the waiting starts just after the pop_task call and
ends just before the running_tasks value is incremented, which would
also mean that somehow even with an emptied list a new task was added
just after), we make sure the host_job is not in a DONE state in the
probe and reply tasks. So that should be good.