docs: update TODO and priorities in FIX list
The ping function does not seem to be essential to handle the RTT
timeout. It will just allow us to get a more accurate value for the
initial RTT timeout. Which is good, but a reasonable default value with
an reactive update function will do about the same. Maybe do it later
when the RTT timeout is handled dynamically and that a congestion window
has been implemented (with slow start and exponential backoff).