feat: compute rtt estimation for probe timeout
This looks like it works, although now ft_nmap is both less precise and
longer to execute... Well this was to be expected. The result needs to
be refined. Like, right now min and max rtt timeout are totally useless,
they should be used as inteded, to limit rtt_timeout.

It seems to be increasing way too fast, which tells me that some packets
are missed or that the locking system (the new rtt_mutex added to the
mutex list) is taking too much time so RTT instance values are
artificially inflated. This is funny, I honestly thought the exact
inverse would occur. Meaning that sometimes, since we cannot
differentiate one probe from an other, if the response to a probe that
just timed out arrived after the retry then rtt_update would think the
rtt was lower than it is. By the way, I still think this problem can
also happen, so this will have to checked out.