feat: add send window to control probe output
This does not really work yet. There is a problem for handling retries
and timeouts with this method (as usual with ft_nmap anyway). Right now
the 'current' value of the window is decreased each time a probe task is
going to be executed, which is wrong because reply probe do not
necessarily correspond to a probe in the window. If the probe has
already been replied to then no packet is going to be sent, so the
value of 'current' should not change.