fix: set a limit to pseudo thread tasks count
The limit has been arbitrarily set to 16. The point of this is to
prevent looping on send tasks for monothreaded mode. The problem was
that at first there is a lot of probe tasks to be executed, so it just
dit them all in a row, which takes time. While doing that it was not
getting incoming probes, so it just looped until a timeout was reached.
Then it was reading every response at the same time and some ports were
apprearing as fitlered when they should not have been.