feat: use a different window for each scan type
This makes it possible to handle timeouts on a scan to scan basis. Because
a certain type of response can be filtered, but not an other. Especially
with UDP. It makes no sense to slow every TCP scan down when only the UDP
is heavily filtered. Likewise for SYN and ACK scan which can have very
different reponse rates.