test: find random ip with shell script
Super cool little script to find random targets to scan. The only problem
is that it is a bit slow. There might be a way to make it much faster and
efficient by using curl instead of ping. It would only return addresses
corresponding to a web sever. Of course it would mean the result would be
less diverse.