feat: execute ping scans for host discovery
This is mainly to estimate the rtt before scanning. By default no host
is skipped because it did not respond to an echo scan. This is because
a lot of hosts actually do not respond to echo requests. The original
nmap uses a lot of different ping techniques at the same time by default.
This would need to be done too to fix this. This is a lot of work, so
just providing a 'skip-non-responsive' option is a faster way to deal
with this.